pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["lynma01 <lynma01@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.8"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
